chore(wizard-refactor): extract tools/create_community
Community-hub scaffold (~230 lines). No pinning dispatcher tests but
covered by sibling integration suites; 61/61 green across wizard tests.
